:hammer: Be less permissive about Tree event handling
The "fix" I did to handle the lack of `sender` on the `Tree` events after
bumping from textual 0.11.0 to 0.22.0 meant I reintroduced a visual bug
where the highlight information from the `DirectoryTree` would leak into the
source path information.

This adds the guard back, but in doing so has to dive into the
`DirectoryTree` internals to make it work.

See https://github.com/Textualize/textual/issues/2413 for some more context.

I think as part of the `DirectoryTree` revamp we're going to do a
double-check of the apparent provenance of events would be a good idea too.
It doesn't make a lot of sense to me that DirectoryTree "leaks" events to
Tree event handlers.