:sparkles: Require a file on the command line now
Now that I've dropped the DirectoryTree as the way to find a file to view,
when opening a fresh one, and used textual-fspicker instead, I think this
makes more sense.

I could move to opening the FileOpen dialog if nothing is passed on the
command line, but I'm not sure I like that idea really.